third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/scroll-snap-type-expected.txt

This is a testharness.js-based test.
[FAIL] 'scroll-snap-type' does not support 'inline proximity'
  assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
Harness: the test ran to completion.
